// apps/api/src/@generated/prisma-nestjs-graphql/job-profile-history/job-profile-history-min-aggregate.input.ts
import { Field, InputType } from '@nestjs/graphql';

@InputType()
export class JobProfileHistoryMinAggregateInput {
  @Field(() => Boolean, { nullable: true })
  id?: true;

  @Field(() => Boolean, { nullable: true })
  job_profile_id?: true;

  @Field(() => Boolean, { nullable: true })
  created_at?: true;

  @Field(() => Boolean, { nullable: true })
  created_by?: true;

  @Field(() => Boolean, { nullable: true })
  updated_at?: true;

  @Field(() => Boolean, { nullable: true })
  updated_by?: true;

  @Field(() => Boolean, { nullable: true })
  deleted_at?: true;

  @Field(() => Boolean, { nullable: true })
  deleted_by?: true;
}